refactor(renderer): Using common function for creating url
