+++ cmd/fibr/templates/async-image.html
@@ -1,7 +1,7 @@
 {{ define "async-image" }}
   <picture id="picture-{{ .File.ID }}" class="no-margin" data-picture data-alt="Thumbnail of {{ .File.Name }}">
     <noscript>
-      <img class="thumbnail full" src="{{ .Config.PublicURL }}/{{ urlquery (.File.Name) }}?thumbnail&v={{ .Version }}" alt="Thumbnail of {{ .File.Name }}" />
+      <img class="thumbnail full" src="{{ url (.Request.URL .File.Name) }}?thumbnail&v={{ .Version }}" alt="Thumbnail of {{ .File.Name }}" />
     </noscript>
   </picture>
 {{ end }}
