+++ cmd/fibr/templates/favicon.html
@@ -1,8 +1,8 @@
 {{ define "favicon" }}
-  <link rel="apple-touch-icon" sizes="180x180" href="{{ .Config.PublicURL }}/favicon/apple-touch-icon.png?v={{ .Config.Version }}">
-  <link rel="icon" type="image/png" sizes="32x32" href="{{ .Config.PublicURL }}/favicon/favicon-32x32.png?v={{ .Config.Version }}">
-  <link rel="icon" type="image/png" sizes="16x16" href="{{ .Config.PublicURL }}/favicon/favicon-16x16.png?v={{ .Config.Version }}">
-  <link rel="manifest" href="{{ .Config.PublicURL }}/favicon/manifest.json?v={{ .Config.Version }}">
-  <link rel="mask-icon" href="{{ .Config.PublicURL }}/favicon/safari-pinned-tab.svg?v={{ .Config.Version }}" color="#5bbad5">
+  <link rel="apple-touch-icon" sizes="180x180" href="{{ url "/favicon/apple-touch-icon.png" }}?v={{ .Config.Version }}">
+  <link rel="icon" type="image/png" sizes="32x32" href="{{ url "/favicon/favicon-32x32.png" }}?v={{ .Config.Version }}">
+  <link rel="icon" type="image/png" sizes="16x16" href="{{ url "/favicon/favicon-16x16.png" }}?v={{ .Config.Version }}">
+  <link rel="manifest" href="{{ url "/favicon/manifest.json" }}?v={{ .Config.Version }}">
+  <link rel="mask-icon" href="{{ url "/favicon/safari-pinned-tab.svg" }}?v={{ .Config.Version }}" color="#5bbad5">
   <meta name="theme-color" content="#ffffff">
 {{ end }}
