+++ cmd/fibr/templates/share-list.html
@@ -69,7 +69,7 @@
             {{ range $key, $value := .Shares }}
               <tr class="share-content">
                 <td>
-                  <a href="{{ $root.Config.PublicURL }}/{{ $value.ID }}/">
+                  <a href="{{ url "/" }}{{ $value.ID }}/">
                     <code>{{ $value.ID }}</code>
                   </a>
                 </td>
@@ -78,13 +78,13 @@
                 </th>
                 <td class="center">
                   {{ if .Edit }}
-                    <img class="icon" src="{{ $root.Config.PublicURL }}/svg/edit?fill=silver" alt="Editable">
+                    <img class="icon" src="{{ url "/svg/edit?fill=silver" }}" alt="Editable">
                   {{ end }}
                   {{ if .Password }}
-                    <img class="icon" src="{{ $root.Config.PublicURL }}/svg/lock?fill=silver" alt="Password protected">
+                    <img class="icon" src="{{ url "/svg/lock?fill=silver" }}" alt="Password protected">
                   {{ end }}
                   {{ if .Duration }}
-                    <img class="icon" src="{{ $root.Config.PublicURL }}/svg/hourglass?fill=silver" alt="Duration">
+                    <img class="icon" src="{{ url "/svg/hourglass?fill=silver" }}" alt="Duration">
                     {{ $value.Duration.Hours }}h
                   {{ end }}
                 </td>
@@ -94,7 +94,7 @@
                     <input type="hidden" name="method" value="DELETE" />
                     <input type="hidden" name="id" value="{{ $value.ID }}" />
                     <button type="submit" onclick="return confirm('Are you sure you want to delete share for {{ $value.Path }}?')" class="button button-icon" alt="Delete">
-                      <img class="icon" src="{{ $root.Config.PublicURL }}/svg/times?fill=silver" alt="Delete">
+                      <img class="icon" src="{{ url "/svg/times?fill=silver" }}" alt="Delete">
                     </button>
                   </form>
                 </td>
