+++ cmd/fibr/templates/sitemap.html
@@ -3,7 +3,7 @@
           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
           xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
     <url>
-      <loc>{{ .PublicURL }}</loc>
+      <loc>{{ publicURL "" }}</loc>
       <changefreq>weekly</changefreq>
       <priority>1.00</priority>
     </url>
