+++ go.mod
@@ -4,7 +4,7 @@ go 1.16
 
 require (
 	github.com/ViBiOh/auth/v2 v2.8.4
-	github.com/ViBiOh/httputils/v4 v4.11.4
+	github.com/ViBiOh/httputils/v4 v4.11.6
 	github.com/prometheus/client_golang v1.10.0
 	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a
 	golang.org/x/text v0.3.6
