+++ pkg/fibr/renderer.go
@@ -24,7 +24,7 @@ func FuncMap(thumbnailApp thumbnail.App) template.FuncMap {
 			}
 		},
 		"rebuildPaths": func(parts []string, index int) string {
-			return path.Join(parts[:index+1]...)
+			return fmt.Sprintf("/%s", path.Join(parts[:index+1]...))
 		},
 		"iconFromExtension": func(file provider.RenderItem) string {
 			extension := file.Extension()
@@ -61,10 +61,6 @@ func (a app) TemplateFunc(w http.ResponseWriter, r *http.Request) (string, int, 
 		return "", 0, nil, model.WrapMethodNotAllowed(errors.New("you lack of method for calling me"))
 	}
 
-	if r.URL.Path == "/sitemap.xml" {
-		return "sitemap", http.StatusOK, nil, nil
-	}
-
 	if query.GetBool(r, "redirect") {
 		a.rendererApp.Redirect(w, r, r.URL.Path, renderer.ParseMessage(r))
 		return "", 0, nil, nil
@@ -73,7 +69,6 @@ func (a app) TemplateFunc(w http.ResponseWriter, r *http.Request) (string, int, 
 	request, err := a.parseRequest(r)
 	if err != nil {
 		if errors.Is(err, model.ErrUnauthorized) {
-			fmt.Println("kiki")
 			w.Header().Add("WWW-Authenticate", `Basic realm="fibr" charset="UTF-8"`)
 		}
 		return "", 0, nil, err
